fix(exams): fix exams command
